refactor ethereum wallet functionality to pkg
