+++ internal/api/handler/rpc_get_challenge_test.go
@@ -2,12 +2,12 @@ package handler
 
 import (
 	"context"
+	"github.com/kyamagames/auth/pkg/util"
 	"testing"
 
 	"github.com/kyamagames/auth/api/pb"
 	mockcache "github.com/kyamagames/auth/internal/cache/mock"
 	mockdb "github.com/kyamagames/auth/internal/db/mock"
-	"github.com/kyamagames/auth/internal/util"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
 )
