+++ internal/api/handler/rpc_verify_access_token_test.go
@@ -2,6 +2,7 @@ package handler
 
 import (
 	"context"
+	"github.com/kyamagames/auth/pkg/util"
 	"testing"
 	"time"
 
@@ -9,7 +10,6 @@ import (
 	mockcache "github.com/kyamagames/auth/internal/cache/mock"
 	mockdb "github.com/kyamagames/auth/internal/db/mock"
 	"github.com/kyamagames/auth/internal/token"
-	"github.com/kyamagames/auth/internal/util"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
 )
