+++ internal/db/sqlc/account.sql_test.go
@@ -2,9 +2,8 @@ package db
 
 import (
 	"context"
+	"github.com/kyamagames/auth/pkg/util"
 	"testing"
-
-	"github.com/kyamagames/auth/internal/util"
 
 	"github.com/stretchr/testify/require"
 )
