+++ internal/token/paseto_maker_test.go
@@ -1,12 +1,12 @@
 package token
 
 import (
+	"github.com/kyamagames/auth/pkg/util"
 	"testing"
 	"time"
 
 	"github.com/brianvoe/gofakeit/v6"
 	"github.com/google/uuid"
-	"github.com/kyamagames/auth/internal/util"
 	"github.com/stretchr/testify/require"
 )
